Add URL option for sampling strategies file (#2519)
* Add url option for sampling strategies

This change will let user to provide a
URL to download sampling strategies. Default
strategy is the fallback option if the URL is
temporarily unavailable.

Signed-off-by: Deepak <sah.sslpu@gmail.com>

* Update NewStrategyStore constructor

This updates the logic for NewStrategyStore
constructor to skip any loading/reloading logic
to execute when StrategiesFile is empty.

Signed-off-by: Deepak <sah.sslpu@gmail.com>

* Refactor strategy store tests

This refactors the strategy store tests, moving tests
with URL to separate test functions.

Signed-off-by: Deepak <sah.sslpu@gmail.com>

* Use strings.Replace to update strategies

Signed-off-by: Deepak <sah.sslpu@gmail.com>

* Change strategiesJSON to a function

This updates the mockStrategyServer to return
a mock strategy along with a mock server. The
returned strategy can be used to update it externally.
Also, changed package-level var strategiesJSON to
a function to return strategies JSON with a given probability.

Signed-off-by: Deepak <sah.sslpu@gmail.com>

* Fix linter failures

Signed-off-by: Deepak <sah.sslpu@gmail.com>